refactor(005): prevent editing learning profile if onboarding is incomplete
